refactor(tests): split cave home migration coverage
