+++ scripts/run-tests.mjs
@@ -839,7 +839,10 @@ export const SUITES = {
     "scripts/beads-pr-bridge.test.mjs",
     "scripts/beads-pr-patrol.test.mjs",
     "src/lib/coven-paths.test.ts",
-    "src/lib/server/cave-home-migration.test.ts",
+    "src/lib/server/cave-home-migration-backup-recovery.test.ts",
+    "src/lib/server/cave-home-migration-json-merge.test.ts",
+    "src/lib/server/cave-home-migration-status-recovery.test.ts",
+    "src/lib/server/cave-home-migration-locks-takeover.test.ts",
     "src/lib/server/cave-home-migration-status.test.ts",
     "src/lib/server/global-npm-install-lane.test.ts",
     "src/app/api/cave-home-migration/route.test.ts",
@@ -1162,7 +1165,10 @@ const ALIAS_LOADER = new Set([
   "src/lib/stitch-patterns.test.ts",
   "src/lib/server/space-usage.test.ts",
   "src/lib/server/stuck-created-sweep.test.ts",
-  "src/lib/server/cave-home-migration.test.ts",
+  "src/lib/server/cave-home-migration-backup-recovery.test.ts",
+  "src/lib/server/cave-home-migration-json-merge.test.ts",
+  "src/lib/server/cave-home-migration-status-recovery.test.ts",
+  "src/lib/server/cave-home-migration-locks-takeover.test.ts",
   "src/lib/server/cave-home-migration-status.test.ts",
   "src/lib/server/preferences-store.test.ts",
   "src/lib/server/backdrop-store.test.ts",
